% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{kmeans}
\alias{kmeans}
\title{Calculate cluster centroids using K-means}
\usage{
kmeans(inputFile, numberClusters, outputFile)
}
\arguments{
\item{inputFile}{path to input file in the OPF binary format}

\item{numberClusters}{number of clusters (k)}

\item{outputFile}{path to output file containing cluster centroids}
}
\description{
Calculate cluster centroids using K-means
}
\details{
Creates a file with the cluster centroids.
}
