% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf2svm}
\alias{opf2svm}
\title{Converts OPF to SVM}
\usage{
opf2svm(inputFile, outputFile)
}
\arguments{
\item{inputFile}{path to OPF input file}

\item{outputFile}{path to SVM output file}
}
\description{
Converts OPF to SVM
}
