% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf2txt}
\alias{opf2txt}
\title{Converts OPF to text}
\usage{
opf2txt(inputFile, outputFile)
}
\arguments{
\item{inputFile}{path to OPF input file}

\item{outputFile}{path to text output file}
}
\description{
Converts OPF to text
}
