% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_check}
\alias{opf_check}
\title{Check the OPF file}
\usage{
opf_check(inputFile)
}
\arguments{
\item{inputFile}{path to input ASCII file in the LibOPF format}
}
\description{
Check the OPF file
}
\details{
usage opf_check <input ASCII file in the LibOPF format>:
Note that the input file for opf_check must be a text file.
Use opf2txt to convert your OPF binary file into a text file.
}
