% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_classify}
\alias{opf_classify}
\title{Executes the test phase of the OPF classifier}
\usage{
opf_classify(dataSet, classifier, precomputedDistance = NA)
}
\arguments{
\item{dataSet}{Testing subgraph object}

\item{classifier}{The classifier model object}

\item{precomputedDistance}{The precomputed distance matrix (leave it in blank if you are not using this resource)}
}
\value{
Returns the given graph classification array
}
\description{
Executes the test phase of the OPF classifier
}
