% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_cluster}
\alias{opf_cluster}
\title{Computes clusters by OPF}
\usage{
opf_cluster(dataSet, kmax, calculateOp, value, precomputedDistance = NA)
}
\arguments{
\item{dataSet}{Unlabeled subgraph object}

\item{kmax}{kmax(maximum degree for the knn graph)}

\item{calculateOp}{how its calculate. 0 for (height), 1 for (area) and 2 for (volume)}

\item{value}{value of parameter "op" in (0-1)}

\item{precomputedDistance}{The precomputed distance subgraph object (leave it in blank if you are not using this resource)}
}
\value{
Returns a list which contains the classifier object and the classificantion array object
}
\description{
Computes clusters by OPF
}
