% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_CreateSubgraph}
\alias{opf_CreateSubgraph}
\title{Creates a empty opf graph structure}
\usage{
opf_CreateSubgraph(nnodes)
}
\arguments{
\item{nnodes}{number of nodes}
}
\value{
Returns the empty opf graph
}
\description{
Creates a empty opf graph structure
}
