% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_distance}
\alias{opf_distance}
\title{Generates the precomputed distance file for the OPF classifier}
\usage{
opf_distance(dataSet, distanceOp, normalize = 0)
}
\arguments{
\item{dataSet}{Subgraph object, normaly is the full data}

\item{distanceOp}{distance calculation option}

\item{normalize}{distance normalization? 1- yes 0 - no}
}
\value{
Returns a distance matrix
}
\description{
Generates the precomputed distance file for the OPF classifier
}
\details{
Options for distance calculation:\cr
1 - Euclidean\cr
2 - Chi-Square\cr
3 - Manhattan (L1)\cr
4 - Canberra\cr
5 - Squared Chord\cr
6 - Squared Chi-Squared\cr
7 - BrayCurtis
}
