% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_info}
\alias{opf_info}
\title{Gives information about the OPF file}
\usage{
opf_info(dataSet)
}
\arguments{
\item{dataSet}{The OPF file}
}
\description{
Gives information about the OPF file
}
\examples{
\dontrun{
opf_info(dataset)
}

}
