% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_knn_classify}
\alias{opf_knn_classify}
\title{Executes the test phase of the OPF classifier with knn adjacency}
\usage{
opf_knn_classify(dataSet, classifier, precomputedDistance = NA)
}
\arguments{
\item{dataSet}{The testing object produced by the opf_split (subGraph object)}

\item{classifier}{The classifier object produced by one of the classification functions (model object)}

\item{precomputedDistance}{The precomputed distance matrix produced by the opf_distance (leave it in blank if you are not using this resource)}
}
\value{
Returns the given subGraph classification list
}
\description{
Executes the test phase of the OPF classifier with knn adjacency
}
