% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_knn_train}
\alias{opf_knn_train}
\title{Executes the training phase of the OPF classifier with knn adjacency}
\usage{
opf_knn_train(trainFile, evaluatFile, kmax, precomputedDistance = NA)
}
\arguments{
\item{trainFile}{Training subgraph object}

\item{evaluatFile}{Evaluation subgraph object}

\item{kmax}{kmax(maximum degree for the knn graph)}

\item{precomputedDistance}{The precomputed distance matrix (leave it in blank if you are not using this resource)}
}
\description{
Executes the training phase of the OPF classifier with knn adjacency
}
\details{
Creates a file .out, which is used in opf_knn_classify, for example. Also creates a .time that contais the processing time.
}
