% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_learn}
\alias{opf_learn}
\title{Executes the learning phase for the OPF classifier}
\usage{
opf_learn(trainFile, evaluatFile, precomputedDistance = NA)
}
\arguments{
\item{trainFile}{The training set object}

\item{evaluatFile}{The evaluation set object}

\item{precomputedDistance}{The precomputed distance subgraph object (leave it in blank if you are not using this resource)}
}
\description{
Executes the learning phase for the OPF classifier
}
\details{
Execute the training phase and generates an output file named classifier.opf
}
