% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_merge}
\alias{opf_merge}
\title{Merge subGraphs}
\usage{
opf_merge(dataSets)
}
\arguments{
\item{dataSets}{array of subGraph objects}
}
\value{
Returns the merged subGraph object
}
\description{
Merge subGraphs
}
