% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_normalize}
\alias{opf_normalize}
\title{Normalizes data for the OPF classifier}
\usage{
opf_normalize(dataSet, normalizedOutput)
}
\arguments{
\item{dataSet}{path to dataSet in the OPF file format}

\item{normalizedOutput}{path to normalized output dataSet in the OPF file format}
}
\description{
Normalizes data for the OPF classifier
}
