% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_pruning}
\alias{opf_pruning}
\title{Executes the pruning algorithm}
\usage{
opf_pruning(
  dataTraining,
  dataEvaluating,
  percentageAccuracy,
  precomputedDistance
)
}
\arguments{
\item{dataTraining}{The training object produced by the opf_split (subGraph object)}

\item{dataEvaluating}{The evaluating object produced by the opf_split (subGraph object)}

\item{percentageAccuracy}{Percentage of accuracy [0,1]}

\item{precomputedDistance}{The precomputed distance matrix produced by the opf_distance (leave it in blank if you are not using this resource)}
}
\value{
Returns the classifier model object
}
\description{
Executes the pruning algorithm
}
