% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_read_subGraph}
\alias{opf_read_subGraph}
\title{Reads a file which contains the subGraph}
\usage{
opf_read_subGraph(file)
}
\arguments{
\item{file}{file name}
}
\value{
Returns the subGraph object
}
\description{
Reads a file which contains the subGraph
}
