% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_ReadClassification}
\alias{opf_ReadClassification}
\title{Reads a file which has the nodes classification}
\usage{
opf_ReadClassification(file)
}
\arguments{
\item{file}{file which has the nodes classification}
}
\value{
Returns the nodes classification array
}
\description{
Reads a file which has the nodes classification
}
