% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_ReadDistances}
\alias{opf_ReadDistances}
\title{Reads a file which has the precalculated distances}
\usage{
opf_ReadDistances(file)
}
\arguments{
\item{file}{file which has the distances matrix}
}
\value{
Returns the precalculated distances matrix
}
\description{
Reads a file which has the precalculated distances
}
