% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_ReadModelFile}
\alias{opf_ReadModelFile}
\title{Reads a file which has the classificator}
\usage{
opf_ReadModelFile(file)
}
\arguments{
\item{file}{file which has the classificator}
}
\value{
Returns the classificator object
}
\description{
Reads a file which has the classificator
}
