% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_ReadSubGraph}
\alias{opf_ReadSubGraph}
\title{Reads a file which has the opf graph}
\usage{
opf_ReadSubGraph(file)
}
\arguments{
\item{file}{file where the graph is}
}
\value{
Returns the graph object
}
\description{
Reads a file which has the opf graph
}
