% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_semi}
\alias{opf_semi}
\title{Executes the semi supervised training phase of the OPF classifier}
\usage{
opf_semi(
  labeledTrainFile,
  unLabeledTrainFile,
  evaluatFile = NA,
  precomputedDistance = NA
)
}
\arguments{
\item{labeledTrainFile}{Labeled training subgraph object}

\item{unLabeledTrainFile}{Unlabeled training subgraph object}

\item{evaluatFile}{Evaluation subgraph object}

\item{precomputedDistance}{The precomputed distance matrix (leave it in blank if you are not using this resource)}
}
\description{
Executes the semi supervised training phase of the OPF classifier
}
\details{
Returns the classifier model object
}
