% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_split}
\alias{opf_split}
\title{Generates training, evaluation and test sets for the OPF classifier}
\usage{
opf_split(dataSet, training_p, evaluating_p, testing_p, normalize)
}
\arguments{
\item{dataSet}{The data (subGraph object)}

\item{training_p}{percentage for the training set size [0,1]}

\item{evaluating_p}{percentage for the evaluation set size [0,1] (leave 0 in the case of no learning)}

\item{testing_p}{percentage for the test set size [0,1]}

\item{normalize}{distance normalization? 1- yes 0 - no}
}
\value{
Returns the Training, Evaluating and the Testing objects
}
\description{
Generates training, evaluation and test sets for the OPF classifier
}
