% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_train}
\alias{opf_train}
\title{Executes the training phase of the OPF classifier}
\usage{
opf_train(dataSet, precomputedDistance = NA)
}
\arguments{
\item{dataSet}{The training subgraph object}

\item{precomputedDistance}{The precomputed distance matrix (leave it in blank if you are not using this resource)}
}
\value{
Returns a list which contains the classifier object and the classificantion array object
}
\description{
Executes the training phase of the OPF classifier
}
