% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_write_classification}
\alias{opf_write_classification}
\title{Writes into a file the predicted labels produced by the opf classificator}
\usage{
opf_write_classification(classes, file)
}
\arguments{
\item{classes}{classification list (i.e.,predicted labels) produced by the classifier}

\item{file}{where you want to save the classification vector}
}
\description{
Writes into a file the predicted labels produced by the opf classificator
}
