% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_write_distances}
\alias{opf_write_distances}
\title{Writes into a file the precalculated distances computed by opf_distances function}
\usage{
opf_write_distances(distances, file)
}
\arguments{
\item{distances}{the matrix produced by the opf distances function}

\item{file}{file name where you want to save the distances}
}
\description{
Writes into a file the precalculated distances computed by opf_distances function
}
