% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_write_modelfile}
\alias{opf_write_modelfile}
\title{Writes into a file the trained OPF classifier}
\usage{
opf_write_modelfile(g, file)
}
\arguments{
\item{g}{the classifier object}

\item{file}{file name to save the classifier}
}
\description{
Writes into a file the trained OPF classifier
}
