% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_WriteClassification}
\alias{opf_WriteClassification}
\title{Writes into a file the node classification vector produced by the opf classificator}
\usage{
opf_WriteClassification(classes, file)
}
\arguments{
\item{classes}{classification vector produced by the classificator
file where you want to save the classification vector}
}
\description{
Writes into a file the node classification vector produced by the opf classificator
}
