% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_WriteDistances}
\alias{opf_WriteDistances}
\title{Writes into a file the precalculated distances of the gives set produced by the opf distances}
\usage{
opf_WriteDistances(distances, fileName)
}
\arguments{
\item{distances}{the matrix produced by the opf distances
file file where you want to save the distances}
}
\description{
Writes into a file the precalculated distances of the gives set produced by the opf distances
}
