% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_WriteModelFile}
\alias{opf_WriteModelFile}
\title{Writes into a opf file the trained opf classifier}
\usage{
opf_WriteModelFile(g, file)
}
\arguments{
\item{g}{the classifier object
file where you want to save the classificator}
}
\description{
Writes into a opf file the trained opf classifier
}
