% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_WriteSubGraph}
\alias{opf_WriteSubGraph}
\title{Writes into a opf file a graph}
\usage{
opf_WriteSubGraph(g, file)
}
\arguments{
\item{g}{the graph object
file file where you want to save the graph}
}
\description{
Writes into a opf file a graph
}
