% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opfknn_classify}
\alias{opfknn_classify}
\title{Executes the test phase of the OPF classifier with knn adjacency}
\usage{
opfknn_classify(dataSet, precomputedDistance = NA)
}
\arguments{
\item{dataSet}{path to test set in the OPF file format}

\item{precomputedDistance}{path to precomputed distance file (leave it in blank if you are not using this resource)}
}
\description{
Executes the test phase of the OPF classifier with knn adjacency
}
\details{
Creates a file .out, which is used in opf_accuracy, for example.
}
