% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opfknn_train}
\alias{opfknn_train}
\title{Executes the training phase of the OPF classifier with knn adjacency}
\usage{
opfknn_train(trainFile, evaluatFile, kmax, precomputedDistance = NA)
}
\arguments{
\item{trainFile}{path to training set in the OPF file format}

\item{evaluatFile}{path to evaluating set in the OPF file format (used to learn k)}

\item{kmax}{kmax(maximum degree for the knn graph)}

\item{precomputedDistance}{path to precomputed distance file (leave it in blank if you are not using this resource)}
}
\description{
Executes the training phase of the OPF classifier with knn adjacency
}
\details{
Creates a file .out, which is used in opf_knn_classify, for example. Also creates a .time that contais the processing time.
}
