% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{statistcs}
\alias{statistcs}
\title{Calculates mean of accuracy}
\usage{
statistcs(inputFile, numberAccuracy, message)
}
\arguments{
\item{inputFile}{path to accuracy file}

\item{numberAccuracy}{how much accuracy will take}

\item{message}{text that will be show}
}
\description{
Calculates mean of accuracy
}
