% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{svm2opf}
\alias{svm2opf}
\title{Converts SVM to OPF}
\usage{
svm2opf(inputFile, outputFile)
}
\arguments{
\item{inputFile}{path to SVM input file}

\item{outputFile}{path to OPF output file}
}
\description{
Converts SVM to OPF
}
