% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{txt2opf}
\alias{txt2opf}
\title{Converts text to OPF}
\usage{
txt2opf(inputFile, outputFile)
}
\arguments{
\item{inputFile}{path to text input file}

\item{outputFile}{path to OPF output file}
}
\description{
Converts text to OPF
}
