% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_accuracy}
\alias{opf_accuracy}
\title{Computes the OPF accuracy}
\usage{
opf_accuracy(dataSet, classification)
}
\arguments{
\item{dataSet}{Data object used in the opf_classify function (subGraph object), normaly is the testing object}

\item{classification}{The output list produced by opf_classify function}
}
\value{
Returns the classification accuracy
}
\description{
Computes the OPF accuracy
}
