% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_accuracy}
\alias{opf_accuracy}
\title{Computes the OPF accuracy of a given set}
\usage{
opf_accuracy(dataSet)
}
\arguments{
\item{dataSet}{path to data set in the OPF file format (.dat and .dat.out)}
}
\description{
Computes the OPF accuracy of a given set
}
\details{
Reads the .dat and .dat.out file with the name that was past in the variable dataSet
Computes and creates a .acc file, with the same name as the input, that contains the accuracy. Also creates a .time that contais the processing time.
}
