% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_accuracy4label}
\alias{opf_accuracy4label}
\title{Computes the OPF accuracy for each class of a given set}
\usage{
opf_accuracy4label(dataSet, classification)
}
\arguments{
\item{dataSet}{Subgraph object, normaly is the training object}

\item{classification}{The classification array produced by the classification of the test subgraph}
}
\value{
Returns the calculated test subgraph classification accuracy for each class
}
\description{
Computes the OPF accuracy for each class of a given set
}
