% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_cluster}
\alias{opf_cluster}
\title{Computes clusters by OPF}
\usage{
opf_cluster(dataSet, kmax, calculateOp, value, precomputedDistance = "")
}
\arguments{
\item{dataSet}{path to unlabeled data set in the OPF file format}

\item{kmax}{kmax(maximum degree for the knn graph)}

\item{calculateOp}{how its calculate. 0 for (height), 1 for (area) and 2 for (volume)}

\item{value}{value of parameter "op" in (0-1)}

\item{precomputedDistance}{path to precomputed distance file (leave it in blank if you are not using this resource}
}
\description{
Computes clusters by OPF
}
\details{
Creates a file .out, which is used in opf_knn_classify, for example.
}
