% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_distance}
\alias{opf_distance}
\title{Generates the precomputed distance file for the OPF classifier}
\usage{
opf_distance(dataSet, distanceOp, normalize)
}
\arguments{
\item{dataSet}{path to dataset in the OPF file format}

\item{distanceOp}{distance calculation option}

\item{normalize}{distance normalization? 1- yes 0 - no}
}
\description{
Generates the precomputed distance file for the OPF classifier
}
\details{
Options for distance calculation:
1 - Euclidean
2 - Chi-Square
3 - Manhattan (L1)
4 - Canberra
5 - Squared Chord
6 - Squared Chi-Squared
7 - BrayCurtis

Reads the dataset and create a distances.dat with the distances.
}
