% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_learn}
\alias{opf_learn}
\title{Executes the learning phase for the OPF classifier}
\usage{
opf_learn(trainFile, evaluatFile, precomputedDistance = "")
}
\arguments{
\item{trainFile}{path to training set in the OPF file format}

\item{evaluatFile}{path to evaluation set in the OPF file format}

\item{precomputedDistance}{path to precomputed distance file (leave it in blank if you are not using this resource}
}
\description{
Executes the learning phase for the OPF classifier
}
\details{
Execute the training phase and generates an output file named classifier.opf
}
