% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_merge}
\alias{opf_merge}
\title{Merge subGraphs}
\usage{
opf_merge(dataSets)
}
\arguments{
\item{dataSets}{An array of subGraph objects}
}
\value{
Returns the merged subGraph object
}
\description{
Merge subGraphs
}
\examples{
\dontrun{
dat <- opf_read_subGraph("C:/boat.dat")
X <- opf_split(dat,0.5,0,0.5,0)
opf_merge(c(X$training,X$testing))
}

}
