% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_normalize}
\alias{opf_normalize}
\title{Normalizes data for the OPF classifier}
\usage{
opf_normalize(dataSet)
}
\arguments{
\item{dataSet}{The subgraph object}
}
\value{
Returns the normalized subgraph
}
\description{
Normalizes data for the OPF classifier
}
