% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_pruning}
\alias{opf_pruning}
\title{Executes the pruning algorithm of the OPF classifier}
\usage{
opf_pruning(
  dataTraining,
  dataEvaluating,
  percentageAccuracy,
  precomputedDistance
)
}
\arguments{
\item{dataTraining}{path to training set in the OPF file format}

\item{dataEvaluating}{path to evaluating set in the OPF file format}

\item{percentageAccuracy}{percentage of accuracy [0,1]}

\item{precomputedDistance}{path to precomputed distance file (leave it in blank if you are not using this resource)}
}
\description{
Executes the pruning algorithm of the OPF classifier
}
