% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_pruning}
\alias{opf_pruning}
\title{Executes the pruning algorithm of the OPF classifier}
\usage{
opf_pruning(
  dataTraining,
  dataEvaluating,
  percentageAccuracy,
  precomputedDistance
)
}
\arguments{
\item{dataTraining}{Training subgraph object}

\item{dataEvaluating}{Evaluating subgraph object}

\item{percentageAccuracy}{percentage of accuracy [0,1]}

\item{precomputedDistance}{The precomputed distance matrix (leave it in blank if you are not using this resource)}
}
\value{
Returns the classifier model object
}
\description{
Executes the pruning algorithm of the OPF classifier
}
