% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_semi}
\alias{opf_semi}
\title{Executes the semi supervised training phase of the OPF classifier}
\usage{
opf_semi(
  labeledTrainFile,
  unLabeledTrainFile,
  evaluatFile = "",
  precomputedDistance = NA
)
}
\arguments{
\item{labeledTrainFile}{path to labeled training set in the OPF file format}

\item{unLabeledTrainFile}{path to unlabeled training set in the OPF file format}

\item{evaluatFile}{path to evaluation set in the OPF file format}

\item{precomputedDistance}{path to precomputed distance file (leave it in blank if you are not using this resource)}
}
\description{
Executes the semi supervised training phase of the OPF classifier
}
\details{
Creates a file .out, which is used in opf_classify, for example.
}
