% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_semi}
\alias{opf_semi}
\title{Executes the semi supervised training phase}
\usage{
opf_semi(
  labeledTrainFile,
  unLabeledTrainFile,
  evaluatFile = NA,
  precomputedDistance = NA
)
}
\arguments{
\item{labeledTrainFile}{The labeled training object (subGraph object)}

\item{unLabeledTrainFile}{The unlabeled training object (subGraph object)}

\item{evaluatFile}{The evaluation object produced by the opf_split (subGraph object)}

\item{precomputedDistance}{The precomputed distance matrix produced by the opf_distance (leave it in blank if you are not using this resource)}
}
\description{
Executes the semi supervised training phase
}
\details{
Returns the learned model object
}
