% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_train}
\alias{opf_train}
\title{Executes the training phase of the OPF classifier}
\usage{
opf_train(dataSet, precomputedDistance = NA)
}
\arguments{
\item{dataSet}{path to training set in the OPF file format}

\item{precomputedDistance}{path to precomputed distance file (leave it in blank if you are not using this resource)}
}
\description{
Executes the training phase of the OPF classifier
}
\details{
Execute the training phase and generates an output file named classifier.opf
}
